refactor: move income to domain module
