+++ src/lib.rs
@@ -6,6 +6,7 @@
 
 pub mod config;
 pub mod db;
+pub mod domain;
 pub mod filter;
 pub mod income;
 pub mod init;
