+++ src/payment/mod.rs
@@ -3,7 +3,7 @@
 //! Reporting calculates amounts to pay for each income, per month or per quarter.
 //! It also tracks payments that are already done.
 
-use crate::income::Income;
+use crate::domain::income::Income;
 
 pub mod report;
