+++ tests/uni_import.rs
@@ -3,11 +3,9 @@
 use std::fs::File;
 
 use chrono::NaiveDateTime;
+use monotax::domain::income::Income;
 use monotax::{
-    income::{
-        criteria::{IncomeCriteria, IncomeCriterion, QuarterFilter},
-        Income,
-    },
+    income::criteria::{IncomeCriteria, IncomeCriterion, QuarterFilter},
     time::Quarter,
     universalbank,
 };
